+++ Packages/com.vareiko.foundation/Editor/Validation/FoundationProjectValidator.cs
@@ -25,6 +25,7 @@ namespace Vareiko.Foundation.Editor.Validation
             "jp.hadashikick.vcontainer",
             "com.cysharp.messagepipe",
             "com.cysharp.messagepipe.vcontainer",
+            "com.unity.nuget.newtonsoft-json",
             "com.unity.inputsystem"
         };